% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataTableToInnerLatex}
\alias{dataTableToInnerLatex}
\title{Convert a data.table into inner part of latex table (table
environment, column names, rules to be produced manually)}
\usage{
dataTableToInnerLatex(dt, outfile = "~/Downloads/test.tex", digits = 0,
  include.rownames = FALSE, include.colnames = FALSE)
}
\arguments{
\item{dt}{data.table to be converted into latex table}

\item{outfile}{path of the destination file}

\item{digits}{digit formating. Useful only if table consists only of numbers
with the same desired formating. If this is not the case, it is easier to
arrange the formating manually on the original data.table.}

\item{include.rownames}{Include rownames? Default is FALSE.}

\item{include.colnames}{Include colnames? Default is FALSE}
}
\value{
Character vector, where each entry corresponds to the row in the
output data table. Main use of this function
is its side product consisting of the 'outfile' latex table.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

