% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_header.R
\name{extract_header_info}
\alias{extract_header_info}
\title{Given a list of objects with regression results, return a list of strings
that serve as header decorations in a latex table}
\usage{
extract_header_info(obj, ...)
}
\arguments{
\item{obj_list}{}
}
\description{
Given a list of objects with regression results, return a list of strings
that serve as header decorations in a latex table
}
\author{
Janko Cizel
}

