% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_table_production.R
\name{extract_selected}
\alias{extract_selected}
\title{Extract desired stats from R estimation object and return result in data.table}
\usage{
extract_selected(obj, select = c("beta", "se", "pval"), digits = 3,
  stars = FALSE)
}
\arguments{
\item{obj}{}

\item{select}{}

\item{digits}{}

\item{stars}{}
}
\value{
data.table
}
\description{
Extract desired stats from R estimation object and return result in data.table
}
\author{
Janko Cizel
}

