% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_table_production.R
\name{header_preparation}
\alias{header_preparation}
\title{Given a list of objects with regression results, return a list of strings
that serve as header decorations in a latex table}
\usage{
header_preparation(obj_list = NULL)
}
\arguments{
\item{obj_list}{}
}
\description{
Given a list of objects with regression results, return a list of strings
that serve as header decorations in a latex table
}
\author{
Janko Cizel
}

