% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_summary_table.R
\name{latextable_basic}
\alias{latextable_basic}
\title{Convert simple data frames into nice latex tables (Primarily useful
for summary stat tables)}
\usage{
latextable_basic(data = NULL, digits = 2, head = TRUE, table.width = 14,
  first.col.width = 3, position = "C", print = TRUE,
  caption = "Test table", label = "test.table", add.rows = NULL,
  notes = "Test note.", textsize = "scriptsize", outfile = NULL)
}
\arguments{
\item{data}{}

\item{digits}{}

\item{head}{}

\item{table.width}{}

\item{position}{}

\item{print}{}

\item{caption}{}

\item{label}{}

\item{add.rows}{}

\item{notes}{}

\item{textsize}{}

\item{outfile}{}
}
\value{
Function primarily used for its side-effect: the latex table
}
\author{
Janko Cizel
}

