% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_labels}
\alias{make_labels}
\title{Make labels from a character vector of variable names}
\usage{
make_labels(vars = NULL, lookup_table = NULL, name_col = NULL,
  label_col = NULL, label_pattern = "[\%s]", split_char = ":")
}
\arguments{
\item{vars}{}

\item{lookup_table}{}

\item{name_col}{}

\item{label_col}{}
}
\value{
character vector of labels, corresponding to the variable names. If
variable name is not available in the lookup table, the returned vector
contains the variable name itself.
}
\description{
Make labels from a character vector of variable names
}
\author{
Janko Cizel
}

