% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_table_production.R
\name{parse_result_list_coef}
\alias{parse_result_list_coef}
\title{Given a list of objects with regression results, return a data.table with
publication-ready formatted results.}
\usage{
parse_result_list_coef(obj_list = NULL, digits = 3, stars = TRUE,
  type = "html")
}
\arguments{
\item{obj_list}{a list of objects with regression results}

\item{digits}{}

\item{stars}{}
}
\description{
Given a list of objects with regression results, return a data.table with
publication-ready formatted results.
}
\author{
Janko Cizel
}

