% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtable.R
\name{rtable}
\alias{rtable}
\title{To be completed}
\usage{
rtable(obj_list, type = "latex", ...)
}
\arguments{
\item{obj_list}{}

\item{type}{}

\item{...}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

