% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_scaffolders.R
\name{scaffoldFigure}
\alias{scaffoldFigure}
\title{Scaffold boilerplate for LaTeX figure}
\usage{
scaffoldFigure(figList = NULL, fig.dir = "./FIG", caption = "Test",
  label = "Test")
}
\arguments{
\item{figList}{}

\item{fig.dir}{}

\item{caption}{}

\item{label}{}
}
\description{
Scaffold boilerplate for LaTeX figure
}
\author{
Janko Cizel
}

