% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_scaffolders.R
\name{scaffoldTable}
\alias{scaffoldTable}
\title{Scaffold LaTeX table from a data.frame}
\usage{
scaffoldTable(data = NULL, type = "latex", header = NULL,
  table.width = 14, digits = 2, position = "C", caption = "Test table",
  label = "test.table", notes = "Test note.", outfile = NULL)
}
\arguments{
\item{data}{}

\item{header}{}

\item{table.width}{}

\item{position}{}

\item{caption}{}

\item{label}{}

\item{notes}{}

\item{outfile}{}
}
\description{
Scaffold LaTeX table from a data.frame
}
\author{
Janko Cizel
}

