% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table_scaffolders.R
\name{htmltable}
\alias{htmltable}
\title{Produce a HTML table from a list of regression objects}
\usage{
htmltable(obj_list = NULL, digits = 3, head = TRUE, stars = TRUE,
  print = TRUE)
}
\arguments{
\item{obj_list}{a list of regression objects}

\item{digits}{number of digits to present in result table}

\item{stars}{}

\item{print}{render the resulting table?}
}
\value{
html table
}
\description{
Produce a HTML table from a list of regression objects
}
\author{
Janko Cizel
}

