% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_table_production.R
\name{parse_result_list_coef}
\alias{parse_result_list_coef}
\title{Given a list of objects with regression results, return a data.table with
publication-ready formatted results.}
\usage{
parse_result_list_coef(obj_list = NULL, digits = 3, stars = TRUE,
  type = "html", coef.presentation = "T1", drop_coef = NULL,
  keep_coef = NULL, order_coef = NULL, label_coef = FALSE,
  lookup_table_coef = NULL, lookup_table_coef_name_col = NULL,
  lookup_table_coef_label_col = NULL, label_pattern = "[\%s]")
}
\arguments{
\item{obj_list}{a list of objects with regression results}

\item{digits}{}

\item{stars}{}
}
\description{
Given a list of objects with regression results, return a data.table with
publication-ready formatted results.
}
\author{
Janko Cizel
}

