% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_table_production.R
\name{parse_result_list_static}
\alias{parse_result_list_static}
\title{Given a list of objects with regression results, return a data.table that
contains `non-coefficient` estimation results}
\usage{
parse_result_list_static(obj_list = NULL, digits = 3, stars = NULL,
  type = "html")
}
\arguments{
\item{obj_list}{}

\item{digits}{}
}
\description{
Given a list of objects with regression results, return a data.table that
contains `non-coefficient` estimation results
}
\author{
Janko Cizel
}

